pyproject.toml
[project]
name = "mimi2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "discord-py>=2.5.2",
    "duckdb>=1.3.2",
    "lancedb>=0.24.2",
    "ollama>=0.5.2",
    "openai>=1.99.1",
    "pandas>=2.3.1",
    "pip>=25.2",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "uuid6>=2025.0.1",
]
